+++ cmd/import_windows.go
@@ -34,5 +34,6 @@ import (
 	_ "github.com/Seagate/cloudfuse/component/libfuse"
 	_ "github.com/Seagate/cloudfuse/component/loopback"
 	_ "github.com/Seagate/cloudfuse/component/s3storage"
+	_ "github.com/Seagate/cloudfuse/component/size_tracker"
 	_ "github.com/Seagate/cloudfuse/component/stream"
 )