+++ component/libfuse/libfuse2_handler.go
@@ -1,5 +1,3 @@
-package libfuse
-
 /*
    Licensed under the MIT License <http://opensource.org/licenses/MIT>.
 
@@ -24,6 +22,8 @@ package libfuse
    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
    SOFTWARE
 */
+
+package libfuse
 
 import (
 	"errors"
@@ -67,6 +67,8 @@ type CgofuseFS struct {
 	// group identifier on linux
 	gid uint32
 }
+
+const windowsDefaultSDDL = "D:P(A;;FA;;;WD)" // Enables everyone on system to have access to mount
 
 // Note: libfuse prepends "/" to the path.
 // TODO: Not sure if this is needed for cgofuse, will need to check
@@ -122,73 +124,44 @@ func (lf *Libfuse) initFuse() error {
 			lf.negativeTimeout)
 
 		// Using SSDL file security option: https://github.com/rclone/rclone/issues/4717
-		// Enables everyone on system to have access to mount
-		options += ",FileSecurity=D:P(A;;FA;;;WD)"
+		windowsSDDL := windowsDefaultSDDL
+		if lf.windowsSDDL != "" {
+			windowsSDDL = lf.windowsSDDL
+		}
+		options += ",FileSecurity=" + windowsSDDL
 	}
 
-	// While reading a file let kernel do readahead for better perf
-	options += fmt.Sprintf(",max_readahead=%d", 4*1024*1024)
-
-	// Max background thread on the fuse layer for high parallelism
-	options += fmt.Sprintf(",max_background=%d", lf.maxFuseThreads)
-
-	if lf.allowOther {
-		options += ",allow_other"
-	}
-	if lf.allowRoot {
-		options += ",allow_root"
-	}
-	if lf.readOnly {
-		options += ",ro"
-	}
-	if lf.nonEmptyMount {
-		options += ",nonempty"
-	}
-
-	if lf.umask != 0 {
-		options += fmt.Sprintf(",umask=%04d", lf.umask)
-	}
-
-	// direct_io option is used to bypass the kernel cache. It disables the use of
-	// page cache (file content cache) in the kernel for the filesystem.
-	if fuseFS.directIO {
-		options += ",direct_io"
-	} else {
-		options += ",kernel_cache"
-	}
+	fuse_options := createFuseOptions(lf.host, lf.allowOther, lf.allowRoot, lf.readOnly, lf.nonEmptyMount, lf.maxFuseThreads, lf.umask)
+	options += fuse_options
 
 	// Setup options as a slice
 	opts := []string{"-o", options}
 
 	// Runs as network file share on Windows only when mounting to drive letter.
 	if runtime.GOOS == "windows" && lf.networkShare && common.IsDriveLetter(lf.mountPath) {
-		// TODO: We can support any type of valid network share path so this path could
-		// be configurable for the config file. But this is a good default.
+		var nameStorage string
 
-		// by default nameStorage will be blank
-		nameStorage := "default"
-		kindStorage := "cloud"
+		serverName, err := os.Hostname()
+		if err != nil {
+			log.Err("Libfuse::initFuse : failed to mount fuse. unable to determine server host name.")
+			return errors.New("failed to mount fuse. unable to determine server host name")
+		}
 		// Borrow bucket-name string from attribute cache
 		if config.IsSet("s3storage.bucket-name") {
-
 			err := config.UnmarshalKey("s3storage.bucket-name", &nameStorage)
 			if err != nil {
-				nameStorage = "default"
+				nameStorage = "s3"
 				log.Err("initFuse : Failed to unmarshal s3storage.bucket-name")
-			} else {
-				kindStorage = "bucket"
 			}
 		} else if config.IsSet("azstorage.container") {
 			err := config.UnmarshalKey("azstorage.container", &nameStorage)
 			if err != nil {
-				nameStorage = "default"
+				nameStorage = "azure"
 				log.Err("initFuse : Failed to unmarshal s3storage.bucket-name")
-			} else {
-				kindStorage = "container"
 			}
 		}
 
-		volumePrefix := fmt.Sprintf("--VolumePrefix=\\%s\\%s", kindStorage, nameStorage)
+		volumePrefix := fmt.Sprintf("--VolumePrefix=\\%s\\%s", serverName, nameStorage)
 		opts = append(opts, volumePrefix)
 	}